+++ src/server_share/mysql_wrapper.cpp
@@ -58,7 +58,7 @@ namespace MSW
 		static string buffer;
 		// reserve space in the buffer according to the mysql documentation
 		buffer.resize(str.size()*2+1);
-		unsigned long resultSize = mysql_escape_string((char*)buffer.data(), str.data(), str.size());
+		unsigned long resultSize = mysql_escape_string((char*)buffer.data(), str.data(), (unsigned long)str.size());
 		
 		// now, resize to the real size
 		buffer.resize(resultSize);
@@ -186,7 +186,7 @@ namespace MSW
 
 		TTime startDate = CTime::getLocalTime();
 
-		int result = mysql_real_query(_MysqlContext, queryString.c_str(), queryString.size());
+		int result = mysql_real_query(_MysqlContext, queryString.c_str(), (unsigned long)queryString.size());
 		if (result != 0)
 		{
 			// in all case, we try to reconnect
@@ -196,7 +196,7 @@ namespace MSW
 				// reconnect and retry the request
 				nlinfo("%p Mysql error errno:%d result:%d : %s, try to reconnect...", _MysqlContext, merrno, result, mysql_error(_MysqlContext));
 				if (_connect())
-					result = mysql_real_query(_MysqlContext, queryString.c_str(), queryString.size());
+					result = mysql_real_query(_MysqlContext, queryString.c_str(), (unsigned long)queryString.size());
 				else
 				{
 					nlwarning("Failed to reopen closed connection to send query '%s'", queryString.c_str());
@@ -280,7 +280,7 @@ namespace MSW
 				myTm.tm_wday = -1;
 				myTm.tm_yday = -1;
 				// Convert the local date into a UTC unix time
-				uint32 t = nl_mktime(&myTm);
+				uint32 t = (uint32)nl_mktime(&myTm);
 
 				time = t;
 			}
@@ -303,7 +303,7 @@ namespace MSW
 				myTm.tm_wday = -1;
 				myTm.tm_yday = -1;
 				// Convert the local date into a UTC unix time
-				uint32 t = nl_mktime(&myTm);
+				uint32 t = (uint32)nl_mktime(&myTm);
 
 				time = t;
 			}
@@ -326,7 +326,7 @@ namespace MSW
 				myTm.tm_wday = -1;
 				myTm.tm_yday = -1;
 				// Convert the local date into a UTC unix time
-				uint32 t = nl_mktime(&myTm);
+				uint32 t = (uint32)nl_mktime(&myTm);
 
 				time = t;
 			}
@@ -349,7 +349,7 @@ namespace MSW
 				myTm.tm_wday = -1;
 				myTm.tm_yday = -1;
 				// Convert the local date into a UTC unix time
-				uint32 t = nl_mktime(&myTm);
+				uint32 t = (uint32)nl_mktime(&myTm);
 
 				time = t;
 			}
